Add page component and minor tweaks to the navbar
