+++ client/src/components/navbar/navbar.jsx
@@ -1,6 +1,6 @@
-import React from "react";
-import { AppBar, Toolbar, Typography, Button } from "@mui/material";
-import { Link, useLocation, useNavigate } from "react-router-dom";
+import React from 'react';
+import { AppBar, Toolbar, Typography, Button, Box } from '@mui/material';
+import { Link, useLocation, useNavigate } from 'react-router-dom';
 import {
   chatPath,
   chatSlug,
@@ -9,86 +9,88 @@ import {
   loginPath,
   profilePath,
   registerPath,
-} from "../../routes/constants";
-import { useCurrentUser } from "../../hooks/useCurrentUser";
-import { userService } from "../../services/user-service";
-import { successToast } from "../../utils/customToast";
+} from '../../routes/constants';
+import { useCurrentUser } from '../../hooks/useCurrentUser';
+import { userService } from '../../services/user-service';
+import { successToast } from '../../utils/customToast';
 
-export const Navbar = () => {
+const AuthenticatedUser = () => {
   const { pathname } = useLocation();
   const user = useCurrentUser();
   const navigate = useNavigate();
 
-  // TODO: get user role and if user is admin, render the clubs approval link too
+  return (
+    <Box>
+      {!pathname.includes(homePath) && (
+        <Button color='inherit' onClick={() => navigate(homePath)}>
+          Browse clubs
+        </Button>
+      )}
+      {!pathname.includes(profilePath) && (
+        <Button color='inherit' onClick={() => navigate(profilePath)}>
+          Profile
+        </Button>
+      )}
 
-  const renderUserOptions = () => {
-    if (user) {
-      return (
-        <>
-          {!pathname.includes(homePath) && (
-            <Button color="inherit" onClick={() => navigate(homePath)}>
-              Browse clubs
-            </Button>
-          )}
-          {!pathname.includes(profilePath) && (
-            <Button color="inherit" onClick={() => navigate(profilePath)}>
-              Profile
-            </Button>
-          )}
+      {!pathname.includes(chatSlug) && (
+        <Button onClick={() => navigate(chatPath)} color='inherit'>
+          Chat
+        </Button>
+      )}
 
-          {!pathname.includes(chatSlug) && (
-            <Button onClick={() => navigate(chatPath)} color="inherit">
-              Chat
-            </Button>
-          )}
+      <Button
+        color='inherit'
+        onClick={() => {
+          navigate(createClubsPath);
+        }}
+      >
+        Create club
+      </Button>
+      <Button
+        color='inherit'
+        onClick={() => {
+          userService.logout(user);
+          successToast('You have been logged out.');
+          navigate(homePath);
+        }}
+      >
+        Logout
+      </Button>
+    </Box>
+  );
+};
 
-          <Button
-            color="inherit"
-            onClick={() => {
-              navigate(createClubsPath);
-            }}
-          >
-            Create club
-          </Button>
-          <Button
-            color="inherit"
-            onClick={() => {
-              userService.logout(user);
-              successToast("You have been logged out.")
-              navigate(homePath)
-            }}
-          >
-            Logout
-          </Button>
-        </>
-      );
-    } else {
-      return (
-        <>
-          <Button onClick={() => navigate(loginPath)} color="inherit">
-            Log in
-          </Button>
-          <Button onClick={() => navigate(registerPath)} color="inherit">
-            Register
-          </Button>
-        </>
-      );
-    }
-  };
+const UnauthenticatedUser = () => {
+  const navigate = useNavigate();
+
+  return (
+    <Box>
+      <Button onClick={() => navigate(loginPath)} color='inherit'>
+        Log in
+      </Button>
+      <Button onClick={() => navigate(registerPath)} color='inherit'>
+        Register
+      </Button>
+    </Box>
+  );
+};
+
+// TODO: get user role and if user is admin, render the clubs approval link too
+export const Navbar = () => {
+  const user = useCurrentUser();
 
   return (
     <AppBar
-      position="static"
-      sx={{ backgroundColor: "#EE7214", fontFamily: "Roboto" }}
+      position='static'
+      sx={{ backgroundColor: '#EE7214', fontFamily: 'Roboto' }}
     >
-      <Toolbar>
-        <Typography variant="h6" component="div" sx={{ flexGrow: 0.5 }}>
-          Book Tennis
-        </Typography>
-        <Button component={Link} to="/" color="inherit">
-          Home
-        </Button>
-        {renderUserOptions()}
+      <Toolbar sx={{ display: 'flex', justifyContent: 'space-around' }}>
+        <Link to='/'>
+          <Typography variant='h6' component='div' sx={{ flexGrow: 0.5 }}>
+            Book Tennis
+          </Typography>
+        </Link>
+        {user ? <AuthenticatedUser /> : <UnauthenticatedUser />}
       </Toolbar>
     </AppBar>
   );
