+++ client/src/routes/constants.js
@@ -1,20 +1,18 @@
 export const root = '/';
 export const loginSlug = 'login';
 export const registerSlug = 'register';
-export const homeSlug = 'home';
+export const browseClubsSlug = 'browse-clubs';
 export const profileSlug = 'profile';
 export const createClubSlug = 'create-club';
 export const chatSlug = 'chat';
 export const clubsApprovalSlug = 'clubs-approval';
-export const dashboardSlug = 'dashboard';
 export const clubSlug = 'club';
 
-export const homePath = `${root}${homeSlug}`;
+export const homePath = `${root}${browseClubsSlug}`;
 export const loginPath = `${root}${loginSlug}`;
 export const registerPath = `${root}${registerSlug}`;
 export const profilePath = `${root}${profileSlug}`;
 export const createClubsPath = `${root}${createClubSlug}`;
 export const chatPath = `${root}${chatSlug}`;
 export const clubsApprovalPath = `${root}${clubsApprovalSlug}`;
-export const dashboardPath = `${root}${dashboardSlug}`;
 export const clubPath = `${root}${clubSlug}`;
