--- client/src/routes/home-page/home-page.jsx
@@ -1,41 +1,0 @@
-import {
-  Box,
-  CssBaseline,
-  Grid,
-  Paper,
-  ThemeProvider,
-  Typography,
-  createTheme,
-} from "@mui/material";
-import "react-toastify/dist/ReactToastify.css";
-import { Navbar } from "../../components/navbar/navbar";
-
-const theme = createTheme();
-
-export const HomePage = () => {
-  // TODO: visualise courts
-
-  return (
-    <ThemeProvider theme={theme}>
-      <Grid container component="main" sx={{ height: "10vh" }}>
-        <CssBaseline />
-        <Navbar/>
-        <Grid item xs={12} component={Paper} elevation={3} square>
-          <Box
-            sx={{
-              my: 6,
-              mx: 4,
-              display: "flex",
-              flexDirection: "column",
-              alignItems: "center",
-            }}
-          >
-            <Typography component="h1" variant="h3" color="#EE7214">
-              Tennis clubs
-            </Typography>
-          </Box>
-        </Grid>
-      </Grid>
-    </ThemeProvider>
-  );
-};
